Churn Analysis Dashboard
This dashboard was built to analyze and visualize the customer churn of a subscription bases service. It was built to help stakeholders better understand the demographics of the most active users and least active users.

Key Findings:
They had lost 20.7% (4,114) of its 19873 users.
They had 9,519 users who active every month.
The Average tenure of users was a little over 5 months.
The demographic of users with the most churn was Old aged people from France and Spain.

Why does this matter?
This research helps stakeholders understand the demographic that would require incentives for using the service to spur them to stay longer and reduce churn.